requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
